+++ bin/otrs.CheckModules.pl
@@ -230,6 +230,12 @@ my @NeededModules = (
         Comment  => 'Required for high resolution timestamps.',
     },
     {
+        # perlcore
+        Module   => 'Time::Piece',
+        Required => 1,
+        Comment  => 'Required for statistics.',
+    },
+    {
         Module   => 'XML::Parser',
         Required => 0,
         Comment  => 'Recommended for faster xml handling.',